Teach ds how to print a simple script
In this commit, I add the --script-mode option, in order to print the
docker commands as a simple script. This way, if I have to run DS on a
machine that lacks even Python, but does support some kind of shell and
docker, I can use this mode to generate the script I'd need to run.